Fix startup crash when replacing map files
Previously, the controller would have crashed if a map was replaced
by another with the exact same file name, since we only checked
the unique UID constraint when inserting; not the unique file name
constraint.

+++ src/controller/facade/on_command.rs
@@ -526,9 +526,11 @@ impl Controller {
 
                 // Delete file, otherwise the map will be scanned back into the
                 // database at the next launch.
-                let map_path = self.config.maps_dir().await.join(map.file_name);
-                if map_path.is_file() {
-                    std::fs::remove_file(map_path).expect("failed to delete map file");
+                if let Some(file_name) = map.file_name {
+                    let map_path = self.config.maps_dir().await.join(file_name);
+                    if map_path.is_file() {
+                        std::fs::remove_file(map_path).expect("failed to delete map file");
+                    }
                 }
 
                 announce(
